+++ src/flight/state/Travel.lua
@@ -1,13 +1,16 @@
-local vehicle = require("abstraction/Vehicle").New()
+local timer      = require("system/Timer").Instance()
+local pub        = require("util/PubSub").Instance()
+local vehicle    = require("abstraction/Vehicle").New()
 local CurrentPos = vehicle.position.Current
+local IsFrozen   = vehicle.player.IsFrozen
 
 ---@class Travel
 ---@field New fun(fsm:FlightFSM):FlightState
 
-local Travel = {}
-Travel.__index = Travel
+local Travel     = {}
+Travel.__index   = Travel
 
-local name = "Travel"
+local name       = "Travel"
 
 ---Creates a new Travel state
 ---@param fsm FlightFSM
@@ -15,10 +18,21 @@ local name = "Travel"
 function Travel.New(fsm)
     local s = {}
 
+    local function openGate()
+        -- When in manual mode, we don't open or close the gate
+        if not IsFrozen() then
+            pub.Publish("SendData", { topic = "GateControl", data = { desiredDoorState = "open" } })
+        end
+    end
+
     function s.Enter()
+        -- Immediately open gate, then repeat evey second.
+        openGate()
+        timer.Add("OpenGate", openGate, 1)
     end
 
     function s.Leave()
+        timer.Remove("OpenGate")
     end
 
     ---Flush